- Nuevos modelos: linea_factura_cliente y linea_factura_proveedor. - Añadida función url() a casi todos los modelos (donde pueda interesar). - Modificados los controladores y vistas que dependen de parametros en la url para obtener un elemento, es decir, general_articulo, general_familia, etc. Ahora ya no muestran errores si no se les pasan parámetros. Por ejemplo, la página general_articulo ya no muestra errores si en la url no se le pasa como parámetro una referencia.
git-svn-id: http://facturascripts.googlecode.com/svn/branches/0.9@345 fe615fe8-962f-0410-811c-9ded47c8d053

+++ controller/admin_pages.php
@@ -32,21 +32,35 @@ class admin_pages extends fs_controller
       {
          if( file_exists('controller/'.$_GET['enable'].'.php') )
          {
-            require_once 'controller/'.$_GET['enable'].'.php';
-            $this->new_fsc = new $_GET['enable']();
-            $this->new_fsc->page->save();
+            if($_GET['enable'] == $this->page->name)
+            {
+               $this->page->save();
+            }
+            else
+            {
+               require_once 'controller/'.$_GET['enable'].'.php';
+               $this->new_fsc = new $_GET['enable']();
+               $this->new_fsc->page->save();
+            }
          }
          else
             $this->new_error_msg("La página no existe");
       }
       else if( isset($_GET['disable']) )
       {
-         $p = new fs_page( array('name'=>$_GET['disable'],
-                                 'title'=>'',
-                                 'folder'=>'',
-                                 'show_on_menu'=>TRUE) );
-         $p->delete();
-         $this->new_error_msg($p->error_msg);
+         if($_GET['disable'] == $this->page->name)
+         {
+            $this->new_error_msg("No puedes desactivar esta página");
+         }
+         else
+         {
+            $p = new fs_page( array('name'=>$_GET['disable'],
+                                    'title'=>'',
+                                    'folder'=>'',
+                                    'show_on_menu'=>TRUE) );
+            $p->delete();
+            $this->new_error_msg($p->error_msg);
+         }
       }
       $this->load_menu();
    }
